ExtractionListenerList: -fixed npe PluginForHost: -updated expectCaptcha and setHasCaptcha, also add accountType in mapID Filer: -updated hasCaptcha -refs thread 62421 build>full
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@40983 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: 58c12efc9a00d0662848c75891aa7afb1b158991

+++ src/org/jdownloader/extensions/extraction/ExtractionListenerList.java
@@ -357,7 +357,7 @@ public class ExtractionListenerList implements ExtractionListener {
             final ArrayList<File> files = new ArrayList<File>(currentArchive.getExtractedFiles());
             files.addAll(currentArchive.getSkippedFiles());
             FileCreationManager.getInstance().getEventSender().fireEvent(new FileCreationEvent(controller, FileCreationEvent.Type.NEW_FILES, files.toArray(new File[files.size()])));
-            if (ex.getSettings().isDeleteInfoFilesAfterExtraction() && currentArchive.getArchiveType().name().startsWith("RAR")) {
+            if (ex.getSettings().isDeleteInfoFilesAfterExtraction() && currentArchive.getArchiveType() != null && currentArchive.getArchiveType().name().startsWith("RAR")) {
                 final File fileOutput = new File(currentArchive.getArchiveFiles().get(0).getFilePath());
                 final File infoFiles = new File(fileOutput.getParentFile(), fileOutput.getName().replaceFirst("(?i)(\\.pa?r?t?\\.?[0-9]+\\.rar|\\.rar)$", "") + ".info");
                 if (infoFiles.exists() && infoFiles.delete()) {
